Extract EventReader
